Moved towards a TREE structure for the original WASM file.  The instructions have fields and bind their operands.
This was a bit challenging because eg: and i32.const instruction doesn't know its the child of an i32.add, we can only know that when the ADD is met, and it "consumes" the two most recent instruction subtrees as its operands, and returns a new instruction stream with the i32.add as the head, replacing its parameters.

The format of the "unit test" text dump is now very messy, but will benefit from the tree structure if appropriately formatted.

The tree structure facilitates some optimisations by making pattern matching MUCH easier, as operands are brought close to the instruction, rather than having a flat list.